Introduce skip flag to avoid rerunning successful builds in later stages

+++ travis/install_linux.sh
@@ -2,4 +2,10 @@
 
 set -e
 
+# Skip this run?
+if [ -f build/skip.txt ]
+then
+  exit 0
+fi
+
 sudo apt-get install -qq -y docker
